Moved a few assert methods to util.TestCaseMixin
